+++ deck/src/perf.rs
@@ -1,7 +1,7 @@
 /// Check if the expression takes too long to execute in debug mode.
 macro_rules! perf {
   ($name:expr, $e:expr, $threshold_ms:expr) => {
-    if cfg!(debug_assertions) {
+    if log::log_enabled!(log::Level::Debug) {
       let start = std::time::Instant::now();
       let res = $e;
       let elapsed = start.elapsed();