+++ deck/scripts/debug.sh
@@ -1,2 +1,2 @@
 #!/bin/bash
-konsole -e "/bin/bash -c '(env RUST_LOG=debug ./launch.sh || true) && read -p \"Press enter to exit\"'"
+konsole -e "/bin/bash -c '(env RUST_LOG=info,stickdeck=debug ./launch.sh || true) && read -p \"Press enter to exit\"'"